requirements.txt
googlesearch-python
requests
bs4